chore: relocate CLI error helpers into plugin CLI
